llm-tracing/requirements.txt
aiohttp==3.11.14
anthropic==0.49.0
fastapi==0.115.11
fastapi[standard]
openai==1.66.3
opentelemetry-api==1.31.0
opentelemetry-exporter-otlp-proto-grpc==1.31.0
opentelemetry-instrumentation==0.52b0
opentelemetry-instrumentation-anthropic==0.38.12
opentelemetry-instrumentation-openai==0.38.12
opentelemetry-proto==1.31.0
opentelemetry-sdk==1.31.0
opentelemetry-semantic-conventions==0.52b0
opentelemetry-semantic-conventions-ai==0.4.3
pydantic==2.10.6
python-dotenv==1.0.1
requests==2.32.3
uvicorn==0.34.0
traceloop-sdk
